New WebConfig for API
