+++ src/main/java/todolist_grupo3/TodolistGrupo3Application.java
@@ -2,28 +2,11 @@ package todolist_grupo3;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @SpringBootApplication
 public class TodolistGrupo3Application {
 
 	public static void main(String[] args) {
 		SpringApplication.run(TodolistGrupo3Application.class, args);
-	}
-
-	@Bean
-	public WebMvcConfigurer corsConfigurer() {
-		return new WebMvcConfigurer() {
-			@Override
-			public void addCorsMappings(CorsRegistry registry) {
-				registry.addMapping("/**")
-					.allowedOrigins("http://localhost:5173")
-					.allowedMethods("GET", "POST", "PUT", "DELETE", "PATCH")
-					.allowedHeaders("*")
-					.allowCredentials(false);
-			}
-		};
 	}
 }
