+++ src/main/java/todolist_grupo3/config/WebConfig.java
@@ -1,0 +1,23 @@
+package todolist_grupo3.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer{
+    @Bean
+	public WebMvcConfigurer corsConfigurer() {
+		return new WebMvcConfigurer() {
+			@Override
+			public void addCorsMappings(CorsRegistry registry) {
+				registry.addMapping("/**")
+					.allowedOrigins("http://localhost:5173")
+					.allowedMethods("GET", "POST", "PUT", "DELETE", "PATCH")
+					.allowedHeaders("*")
+					.allowCredentials(false);
+			}
+		};
+	}
+}
